pyproject.toml
[tool.poetry]
name = "etl-weather-project"
version = "0.1.0"
description = ""
authors = ["Alysson Possebon <apossebon@wiley.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.26.0"
apache-airflow = ">2.10.2,<3.0.0"
psycopg2-binary = "^2.9.10"
pymupdf = "^1.24.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
